Unify `instances` reducers
